+++ CoreStarter/CoreStarter.Api/Controllers/FooController.cs
@@ -1,11 +1,7 @@
-﻿using CoreStarter.Api.Binders;
-using CoreStarter.Api.Configuration;
-using CoreStarter.Api.Models;
+﻿using CoreStarter.Api.Models;
 using CoreStarter.Api.Services;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -19,23 +15,16 @@ namespace CoreStarter.Api.Controllers
     [Route("api/[controller]")]
     public class FooController : ControllerBase
     {
-        private readonly ConnectionStrings _connectionStrings;
         private readonly IFooService _service;
         private readonly ILogger _logger;
 
         /// <summary>
         /// Creates new instance of <see cref="FooController"/>.
         /// </summary>
-        /// <param name="connectionStrings">
-        /// Instance of <see cref="IOptionsSnapshot{ConnectionStrings}"/> object that contains connection string.
-        /// More information: https://docs.microsoft.com/en-us/dotnet/api/microsoft.extensions.options.ioptionssnapshot-1?view=aspnetcore-2.1
-        /// TODO: https://www.strathweb.com/2016/09/strongly-typed-configuration-in-asp-net-core-without-ioptionst/
-        /// </param>
         /// <param name="service">Instance of <see cref="IFooService"/></param>
         /// <param name="logger"></param>
-        public FooController(IOptionsSnapshot<ConnectionStrings> connectionStrings, IFooService service, ILogger<FooController> logger)
+        public FooController(IFooService service, ILogger<FooController> logger)
         {
-            _connectionStrings = connectionStrings.Value ?? throw new ArgumentNullException(nameof(connectionStrings));
             _service = service ?? throw new ArgumentNullException(nameof(service));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         }
@@ -49,7 +38,6 @@ namespace CoreStarter.Api.Controllers
         [HttpPost]
         [ProducesResponseType(typeof(int), 201)]
         [ProducesResponseType(500)]
-        //[SwaggerRequestExample(typeof(Foo), typeof(FooRequestExample))]
         public async Task<IActionResult> Post([FromBody] Foo foo)
         {
             var response = await _service.Create(foo);
@@ -66,16 +54,13 @@ namespace CoreStarter.Api.Controllers
         [HttpPost("content")]
         [ProducesResponseType(typeof(int), 201)]
         [ProducesResponseType(500)]
-        //[AddSwaggerFileUploadButton]
-        // Since we are using custom model provider this post method doesn't support swagger request examples
-        // I guess this can be coded to be supported, but I feel it will go beyond this template's boundaries.
-        public async Task<IActionResult> PostFile([ModelBinder(BinderType = typeof(JsonModelBinder))] Foo foo, IFormFile file)
+        public async Task<IActionResult> PostFile([FromForm] FooFile foo)
         {
             using (var memoryStream = new MemoryStream())
             {
-                await file.OpenReadStream().CopyToAsync(memoryStream);
-
-                var path = Path.Combine(Path.GetTempPath(), file.FileName);
+                await foo.File.OpenReadStream().CopyToAsync(memoryStream);
+                var fileName = Guid.NewGuid().ToString("N");
+                var path = Path.Combine(Path.GetTempPath(), fileName);
                 await System.IO.File.WriteAllBytesAsync(path, memoryStream.ToArray());
             }
 
@@ -91,14 +76,23 @@ namespace CoreStarter.Api.Controllers
         /// <response code="500">Internal server error.</response>
         [HttpGet, ResponseCache(CacheProfileName = "default")]
         [ProducesResponseType(typeof(IEnumerable<Foo>), 200)]
-        //[SwaggerResponseExample(200, typeof(FooResponseExample))]
         public async Task<IActionResult> Get()
         {
-            var response = await _service.Get().ConfigureAwait(false);
-            // Example on how to get configuration values
-            var connections = new List<string> { _connectionStrings.ApiDb, _connectionStrings.Api2Db };
+            var response = await _service.Get();
 
-            return Ok(connections);
+            return Ok(response);
+        }
+
+
+        [HttpPost("throw")]
+        [ProducesResponseType(500)]
+        public async Task<IActionResult> Throw([FromBody] Foo foo)
+        {
+            _ = foo ?? throw new ArgumentNullException(nameof(foo));
+
+            _service.Throw();
+
+            return Ok(foo);
         }
 
         /// <summary>
@@ -111,10 +105,9 @@ namespace CoreStarter.Api.Controllers
         [HttpGet("{id:int:min(1)}", Name = "getById")]
         [ProducesResponseType(typeof(Foo), 200)]
         [ProducesResponseType(404)]
-        //[SwaggerResponseExample(200, typeof(FooListResponseExample))]
         public async Task<IActionResult> Get(int id)
         {
-            var response = await _service.Get(id).ConfigureAwait(false);
+            var response = await _service.Get(id);
 
             if (response == null)
                 return NotFound(id);
@@ -128,11 +121,10 @@ namespace CoreStarter.Api.Controllers
         /// <param name="foo">Instance of <see cref="Foo"/> that holds values that we want updated.</param>
         /// <response code="200">Foo updated successfully.</response>
         /// <response code="500">Internal server error.</response>
-        [HttpPatch]
-        //[SwaggerRequestExample(typeof(Foo), typeof(FooRequestExample))]
+        [HttpPut]
         public async Task<IActionResult> Patch([FromBody] Foo foo)
         {
-            await _service.Update(foo).ConfigureAwait(false);
+            await _service.Update(foo);
 
             return Ok();
         }
@@ -146,7 +138,7 @@ namespace CoreStarter.Api.Controllers
         [HttpDelete("{id:int:min(1)}")]
         public async Task<IActionResult> Delete(int id)
         {
-            await _service.Delete(id).ConfigureAwait(false);
+            await _service.Delete(id);
 
             return Ok();
         }